Reshuffle protocol loading/importing parts
