--- src/utils/importer/index.js
@@ -1,2 +1,0 @@
-export { default as filesystem } from './filesystem';
-export { default } from './importer';
