+++ src/utils/protocol/assetUrl.js
@@ -1,5 +1,5 @@
 /* eslint-disable import/no-mutable-exports */
-import { isElectron } from '../../utils/Environment';
+import { isElectron } from '../Environment';
 
 let assetUrl = null;
 
@@ -7,10 +7,7 @@ if (isElectron()) {
   const electron = window.require('electron');
   const path = electron.remote.require('path');
 
-  assetUrl = (protocolName, assetPath) => {
-    const basename = path.basename(protocolName);
-    return `asset://${path.join(basename, 'assets', assetPath)}`;
-  };
+  assetUrl = (protocolName, assetPath) => `asset://${path.join(protocolName, 'assets', assetPath)}`;
 }
 
 export default assetUrl;
