+++ src/utils/protocol/index.js
@@ -1,3 +1,5 @@
 /* eslint-disable import/prefer-default-export */
 
 export { default as assetUrl } from './assetUrl';
+export { default as protocolPath } from './protocolPath';
+export { default as getProtocol } from './getProtocol';
